Fixed Speed Problems (Read Description
Fixed a major problem for both the Korat and Bullets if a Korat or bullet was faster than than the Korat or bullet in front of it and passed it then its vector positon would not change, changed the code to check the X positon of the first thing in each lane and the thing behind it, movement direction approprate, and swap their positons in the vector if the thing behind the front moved past it to become the new front. Shared pointers saved my ass with a built in function.