loss screen works, removed need for GameViewMenu, migrated audio to the GameViewPlayer class
Made GameViewPlayer have methods for the menuscreen, playscreen, and lossscreen compared to them being separate classes and the game state manager just switches between these method calls based on the state now. Also made it so all assets load at the beginning and audio is now stored and handled by the GameViewPlayer class compared to logic. And also cleaned up the CurrentKoratBullets in the ClearAssets() method of game logic